+++ WaveRecorder/App/Protocols.swift
@@ -1,0 +1,10 @@
+//
+//  Protocols.swift
+//  WaveRecorder
+//
+//  Created by Andrew Steellson on 27.12.2023.
+//
+
+import Foundation
+
+
