+++ WaveRecorder/Services/AudioService.swift
@@ -8,11 +8,19 @@
 import AVFoundation
 
 
-//MARK: - Protocol
+//MARK: - Protocols
 
 protocol AudioServiceProtocol: AnyObject {
     func play(record: Record, onTime time: Float)
     func stop()
+}
+
+protocol AudioServiceRepresentative: AnyObject {
+    func goBack()
+    func play(atTime time: Float, completion: @escaping () -> Void)
+    func stop(completion: @escaping () -> Void)
+    func goForward()
+    func deleteRecord()
 }
